Rename: postboxinfo 파일 componets 폴더로 이동
